requirements.txt
streamlit==1.47.1
python-dotenv==1.1.1
PyPDF2==3.0.1

faiss-cpu==1.12.0
numpy>=1.25,<2.0

reportlab==4.4.3
markdown2==2.5.4
youtube-transcript-api==1.2.2

groq==0.30.0
sentence-transformers==2.7.0
transformers==4.54.1

langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-groq==0.1.9
